inserita un cotrollo per non mostrare i partecipanti che non hanno svolto una verifica

+++ MBSR/startbootstrap-sb-admin-2-gh-pages/tableScore.php
@@ -25,7 +25,7 @@
     	<div class="table-responsive">
         	<table class="table table-bordered" id="dataTable" width="100%" cellspacing="0">
             	<thead>
-                    <tr>
+                    <tr style="background:radial-gradient(circle, rgba(0,149,255,0.06) 2%, rgba(0,149,255,0.08) 100%)">
                       
                       <th>Id</th>
                       <th>Edizione</th>
@@ -42,7 +42,7 @@
                     </tr>
                   </thead>
                   <tfoot>
-                    <tr>
+                    <tr style="background:radial-gradient(circle, rgba(0,149,255,0.06) 2%, rgba(0,149,255,0.08) 100%)">
                     
                     <th>Id</th>
                     <th>Edizione</th>
@@ -72,20 +72,30 @@ if ($result->num_rows > 0){
             
             $idScoring=$row['Id'];
         for ($t = 0; $t < 2; $t++) {
-            $tempo=($t==0 ? 'Pre-test' : 'Post-test' );
-        
-            echo "<tr>
-                   
-                    <td>".$row['Id']."</td>
-                    <td>".$row['edizione']."</td>
-                    <td>".$tempo."</td>
-                    <td>".$row['Cognome']."</td>
-                    <td>".$row['Genere']."</td>";
-            foreach ($score[$t][$idScoring] as $value){
-                            echo "<td>".number_format($value,3,',', ' ')."</td>";
-                        }
-                
-         echo      "</tr>";
+            if (array_key_exists( $idScoring, $score[$t]) ) {
+                $tPiu=$t+1;
+                if ($t==0 && array_key_exists( $idScoring, $score[$tPiu])== false  ) {
+                    $bgcolor=' class= "bg-gradient-warning"';
+                    
+                }else {
+                    $bgcolor=($t==0 ? 'style="background-color:rgb(0,191,255,0.08);"' : 'style="background-color: rgb(255,142,32, 0.08)"' );
+                }
+            
+                $tempo=($t==0 ? 'Pre-test' : 'Post-test' );
+            
+                echo "<tr " .$bgcolor." >
+                       
+                        <td>".$row['Id']."</td>
+                        <td>".$row['edizione']."</td>
+                        <td>".$tempo."</td>
+                        <td>".$row['Cognome']."</td>
+                        <td>".$row['Genere']."</td>";
+                foreach ($score[$t][$idScoring] as $value){
+                                echo "<td>".number_format($value,3,',', ' ')."</td>";
+                            }
+                    
+             echo      "</tr>";
+            }
         }
         }
